Add Link class to the controller
Small refactor in switch module
